+++ src/commands/index.ts
@@ -10,6 +10,7 @@ import editor from './openEditor'
 import review from './review'
 import deepl from './deepl'
 import gotoRange from './gotoRange'
+import gotoNextUsage from './gotoNextUsage'
 import batchHardStringsExtract from './extractStringBulk'
 import { ExtensionModule } from '~/modules'
 
@@ -27,6 +28,7 @@ const m: ExtensionModule = (ctx) => {
     review(ctx),
     deepl(ctx),
     gotoRange(ctx),
+    gotoNextUsage(ctx),
   ])
 }