+++ src/extraction/parsers/html.ts
@@ -31,9 +31,13 @@ export function detect(
   input = input.replace(/<(.*?)={(.*?)}(.*?)>/g, '<$1="$2"$3>')
 
   let lastTag = ''
+  let lastScriptIndex: number | null = null
   const parser = new Parser({
     onopentag(name, attrs) {
       lastTag = name
+      if (name === 'script')
+        lastScriptIndex = parser.endIndex! + 1
+
       if (IGNORED_TAGS.includes(name))
         return
 
@@ -84,17 +88,23 @@ export function detect(
         })
       }
     },
-    ontext(fullText) {
-      const start = parser.startIndex
-      const end = parser.endIndex! + 1
-
-      if (extractScripts && lastTag === 'script') {
-        detections.push(...shiftDetectionPosition(extractScripts(fullText, start), start))
+    onclosetag(name) {
+      if (name !== 'script' || lastScriptIndex == null)
         return
-      }
-
+      const start = lastScriptIndex
+      const fullText = input.slice(start, parser.startIndex!)
+      if (extractScripts)
+        detections.push(...shiftDetectionPosition(extractScripts(fullText, start), start))
+      lastScriptIndex = null
+    },
+    ontext(fullText) {
+      if (lastScriptIndex != null)
+        return
       if (IGNORED_TAGS.includes(lastTag))
         return
+
+      const start = parser.startIndex
+      const end = parser.endIndex! + 1
 
       const text = fullText.split(/\n/g).map(i => i.trim()).filter(Boolean).join(' ')