+++ src/frameworks/fluent-vue-sfc.ts
@@ -1,0 +1,26 @@
+import { Framework } from './base'
+
+class FluentVueSFCFramework extends Framework {
+  id = 'fluent-vue-sfc'
+  display = 'fluent-vue SFC'
+
+  detection = {
+    packageJSON: [
+      'fluent-vue-loader',
+    ],
+  }
+
+  languageIds = []
+
+  usageMatchRegex = []
+
+  refactorTemplates() {
+    return []
+  }
+
+  enableFeatures = {
+    FluentVueSfc: true,
+  }
+}
+
+export default FluentVueSFCFramework