+++ src/views/providers/ProgressProvider.ts
@@ -15,8 +15,12 @@ export class ProgressProvider implements TreeDataProvider<BaseTreeItem> {
   constructor(private ctx: ExtensionContext) {
     const throttledRefresh = throttle(() => this.refresh(), THROTTLE_DELAY)
     EditorPanel.onDidChange(throttledRefresh)
-    CurrentFile.onInitialized(() => CurrentFile.loader.onDidChange(throttledRefresh))
-    Global.onDidChangeLoader(() => Global.loader.onDidChange(throttledRefresh))
+    Global.onDidChangeLoader(() => {
+      throttledRefresh()
+      Global.loader.onDidChange(throttledRefresh)
+      CurrentFile.loader.onDidChange(throttledRefresh)
+    })
+    Global.onDidChangeEnabled(throttledRefresh)
   }
 
   refresh(): void {